requirements.txt
Flask==1.0.2
Jinja2==2.10
jsonschema==3.0.1
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
Pillow==5.4.1
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
matplotlib==3.0.2
waitress==1.2.1
